Account for new "codename" term in error
epgsql 3.2 introduced [a codename atom][1] in the error tuple, thereby
increasing the number of fields, breaking the case statement. Since we
don't particularly care for the specific information returned in this
new field, we can just increase the number of ignored terms to make the
arity match and continue on.

[1]: https://github.com/epgsql/epgsql/commit/8f766ca37ef2f810a2c27bedcbed01b551893e5a